pyproject.toml
[build-system]
requires = ["setuptools >= 74.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hundredandtenserverless"
version = "0.0.0"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Seamus Lowry" },
]
classifiers = [
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "azure-functions==1.24.0",
    "cachecontrol==0.14.4",
    "google-auth==2.48.0",
    "hundredandten==2.1.1",
    "pymongo==4.16.0",
    "requests==2.32.5",
]

[project.optional-dependencies]
lint = [
    "pylint==4.0.5",
    "pyright==1.1.408",
    "ruff==0.15.2",
    "black==26.1.0",
]
test = [
    "coverage==7.13.4",
    "pytest==9.0.2",
]

[tool.coverage.run]
omit = [
    "tests/*"
]

[tool.pyright]
useLibraryCodeForTypes = true

[tool.isort]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]


[tool.ruff]
target-version = "py313"
line-length = 100
lint.extend-select = ["I"]